Switch Regex to a Set for special char check.
Reduces allocations since we aren't creating the substring to use in
a regex match and instead just check for inclusion in a set of chars.

Setup is now also more amenable to extensions since we can update the
parser's special char field `.maybe_special` without having to change
a global regex constant.