Improve types of Parser.[rules,priorities]
This makes `rules` a concrete type, allowing things like
`length(rules)` to become inferrable. Of course the elements are
still not inferrable, but just knowing the container type gets rid of
quite a few inference triggers from running FranklinParser's tests.

This also changes `priorities` to be an `IdDict` rather than a `Dict`.
An `IdDict` is a better choice when each key is of a different type,
because it uses a different strategy for hashing and heavily
uses `@nospecialize`.

These changes shave about 0.8s off the time to run your tests.