'pluggable' string macro parsers.
Possible solution to make it easier to swap/pick the Parser used in a
`cm_str` macro call by using the suffix syntax.

Provides a default 'jmd' syntax that closely matches Markdown.jl and a
'cm' that matches the standard commonmark spec. User-defined parsers can
be used by defining a function in the same module-scope that returns a
`Parser` and then suffixing it's name to the macro call.