Fix interpolation with repeating, mutating names
A variable that is interpolated multiple times with updates in between
will print with the final value in all positions:

```
julia> let x = 1
           function f!(); x += 1; 42; end
           cm"$(x), $(f!()), $(x)"
       end
 2, 42, 2
```

This is because `JuliaValue` is a struct and therefore the two instance
of `JuliaValue(:x, nothing)` are identical. This confuses the
replacement `Dict` in `_interp!`, where value from the earlier position
gets overwritten.

Note: this only affects the interpolation of "pure" variables, since
`Expr` objects are mutable, and so won't match even if they have
identical contents.

This patch introduces a new JuliaExpression type that the parser will
use to add the interpolated expressions into the tree, and it contains a
.pos field to keep track of which expression it is, distinguishing it
from earlier ones. After the evaluation of the interpolated expressions,
these JuliaExpression elements get replaces by the old JuliaValue,
discarding the position information, since that is no longer important.